Added casts to address signedness/precision build errors.
If this code base were being actively maintained, it may be prudent to
reconsider the underlying types.

+++ Vesper/Classes/VSConstants.m
@@ -130,7 +130,7 @@ void VSDefaultsSetTextWeight(VSTextWeight textWeight) {
 	
 	if (textWeight < VSTextWeightRegular || textWeight > VSTextWeightLight)
 		textWeight = VSTextWeightRegular;
-	[[NSUserDefaults standardUserDefaults] setInteger:textWeight forKey:VSDefaultsTextWeightKey];
+	[[NSUserDefaults standardUserDefaults] setInteger:(NSInteger)textWeight forKey:VSDefaultsTextWeightKey];
 	//	[[NSUserDefaults standardUserDefaults] synchronize];
 }
